Extract link activated policy handler to a new file.
